==> 1_3_Fitting.ipynb <==
## Fitting drug response curves with sigmoid function

import pandas as pd
import numpy as np
from tqdm import tqdm
import warnings
warnings.filterwarnings("ignore")
import os, sys
sys.path.insert(1, os.path.relpath("functions"))
from fitting import *
from plotting import *
R2_limit = 0.99

_FOLDER = "results/"
_FOLDER_2 = "data/"

## Fitting data

df_auc = pd.read_csv(_FOLDER+"filt_auc_02.csv")
df_1234 = pd.read_csv(_FOLDER+"filt_1234_02.csv")
drug_curves = pd.read_csv(_FOLDER_2+"normalised_dose_response_data.csv")
conc_columns= ["fd_num_"+str(i) for i in range(10)]
response_norm = ['norm_cells_'+str(i) for i in range(10)]

df_auc.shape, df_1234.shape, drug_curves.shape

functions = [
    "fsigmoid", 
    "sigmoid_2_param",
    "sigmoid_3_param",
    "sigmoid_4_param",
    "logistic_4_param",
    "ll4_4_param",
    "ll4R_4_param",
    "logLogist_3_param"
]

def compare_fitting(df_raw, fitting_functions_list):
    df_results = pd.DataFrame(columns= ["R2>0.9", "R2>0.95", "R2>0.99"])
    for fitting_function in fitting_functions_list:
        df = df_raw.copy()
        print("\n", fitting_function)
        r2, fit_param = fitting_column(df, df.index, x_columns=conc_columns, y_columns= response_norm,
                               fitting_function = fitting_function, default_param=True)
        df[fitting_function+"_r2"] = r2
        df[fitting_function] = fit_param
#         df= df[df[fitting_function+"_r2"]>0]
        print("R2>0:", df.shape)
        print("R2>", R2_limit, df[df[fitting_function+"_r2"]>R2_limit].shape[0])
        df_results.loc[fitting_function, "R2>0.9"] = df[df[fitting_function+"_r2"]>0.9].shape[0]
        df_results.loc[fitting_function, "R2>0.95"] = df[df[fitting_function+"_r2"]>0.95].shape[0]
        df_results.loc[fitting_function, "R2>0.99"] = df[df[fitting_function+"_r2"]>0.99].shape[0]
        print("Number of samples with fitting <0.1:", df[df[fitting_function+"_r2"]<0.1].shape[0])
        print("")
    return df, df_results

%%time
df, df_results = compare_fitting(df_auc, functions)
df_results.to_csv(_FOLDER+"fit_auc_02_compare.csv", index=False)
df.to_csv(_FOLDER+"filt_auc_02_fit.csv", index=False)

%%time
df, df_results_2 = compare_fitting(df_1234, functions)
df_results_2.to_csv(_FOLDER+"fit_1234_compare.csv", index=False)
df.to_csv(_FOLDER+"filt_1234_fit.csv", index=False)

%%time
df, df_results_3 = compare_fitting(drug_curves, functions)
df_results_3.to_csv(_FOLDER+"fit_no_filt_compare.csv", index=False)
df.to_csv(_FOLDER+"filt_fit.csv", index=False)

df_results

df_results.sort_values("R2>0.99")

df_results_2

df_results_2.sort_values("R2>0.99")

df_results_3

df_results_3.sort_values("R2>0.99")
